requirements.txt
attrs==19.3.0
autopep8==1.5.1
bcrypt==3.1.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
dataclasses==0.7
dotenv-python==0.0.1
fastapi==0.54.1
h11==0.9.0
httptools==0.1.1
idna==2.9
importlib-metadata==1.6.0
mongoengine==0.19.1
mongomock==3.19.0
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
py==1.8.1
pycodestyle==2.5.0
pycparser==2.20
pydantic==1.4
PyJWT==1.7.1
pymongo==3.10.1
pyparsing==2.4.7
pytest==5.4.1
python-dotenv==0.12.0
requests==2.23.0
sentinels==1.0.0
six==1.14.0
starlette==0.13.2
urllib3==1.25.8
uvicorn==0.11.3
uvloop==0.14.0
wcwidth==0.1.9
websockets==8.1
zipp==3.1.0
